requirements.txt
pytest
assertpy
allure-pytest
requests
selenium
webdriver_manager
pydantic
pytest-html
